requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
openai==2.15.0
pydantic==2.5.3
python-dotenv==1.0.1
